WIP, first pass at abstracting out TF functionality into a TFAlphaZeroProblem class (#45)
* fix run script with broken import

* gym should work now, but failing on pickling error

* fixed pickling error by moving the gym class definitions outside of the construct_problem function;  TFAZ now runs although i havent verified whether the problem can be solved or not

* comments and light docstrings added

* fix indentation

* added np.all to equals
